+++ src/interfaces/api.ts
@@ -1,4 +1,4 @@
-import { OltInfoItem } from "./OltInfoItem";
+import { Olt } from "./olt";
 
 export interface propsApi{
     city: string;
@@ -8,5 +8,5 @@ export interface propsApi{
     handleError: (error: string) => void;
     startLoading: () => void;
     stopLoading: () => void;
-    OltInfo: OltInfoItem[];
+    OltInfo: Olt[];
 }
